[BUGFIX] Don't install husky on Composer non-git install
Adds a cross-platform condition in the package.json file that disables husky if it isn't needed. When installing the extension through Composer and it is downloaded as archive (rather than using a git clone), it will be downloaded without the .git repository. This will make husky complain because it explicitly requires to be run in a folder containing a folder named ".git". That makes sense, because husky is supposed to add git hooks. If there is no git, you don't need husky either. Arguably, husky should deal with these kind of edge cases, but making it aware of Composer is maybe a bit much to ask.