[FEATURE] Don't render unused sections in StartPage.html
Introduces TypoScript `lib.contentRederCount`. This is used in a condition to test for the existence of content elements with colPos in the `colPos` argument. Compared the `lib.contentRender`, `colPos` is a commaseparated list of colPos values.

`<f:if condition="{f:cObject(typoscriptObjectPath:'lib.contentRenderCount', data:{colPos:'3301,3302,3303'})}">Only rendered if there are content elements in colPos 3301,3302,3303 on the page.</f:if>`